phase3/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlmodel>=0.0.22
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
openai==1.54.4
httpx==0.27.2
pydantic>=2.9.0
